requirements.txt
selenium~=4.8
webdriver-manager~=3.8
pylint==2.11.1
requests==2.26.0
urllib3==1.26.4
jsonpath_rw==1.4.0
jsonpath_rw_ext==1.2.2
pandas==1.4.2
behave==1.2.6
allure-behave==2.13.1


